Refactor extra-properties Admin API: pure-Core services + bundle bridge
Centralizes the extra-property Admin API code and tightens responsibilities:

- The read/write services and the list-record collector move into
  Core\ExtraProperty\Api and become pure Core (repository/reader/writer/
  ShopContext/collector only). Their interfaces are dropped — the single
  consumer uses the concrete services directly.
- ExtraPropertyApiSubscriber moves to PrestaShopBundle\EventListener\API (the
  conventional location) and is the single API-Platform bridge: it resolves
  the entity id (resolveId inlined, replacing the removed ApiResourceIdResolver)
  and converts LANG values locale <-> id_lang via LocalizedValueUpdater, so the
  services it delegates to stay framework-agnostic.
- Validation lives entirely in ExtraPropertyCQRSApiValidator; the payload
  handler is persist-only.
- The response injector no longer strips null values (extra properties are
  dynamic and null is a valid value for a nullable field) and no longer
  re-flattens SHOP scope (the reader already returns a single scalar for the
  active shop constraint); isLangMultishop is derived from the entity.
- ExtraPropertyEndpointTest self-heals: it uninstalls any leftover module
  before install and cleans up unconditionally on teardown, so it can never
  leak the test module into the shared test modules directory.
